Refactor mmc function and main for clarity

+++ Tarefa de Laboratório 05/Q3.c
@@ -1,10 +1,9 @@
-/* gcc -std=c89 -pedantic-errors -Wall -Wextra arquivo.c -o programa */
 #include <stdio.h>
 #include <limits.h>
 
 unsigned long int mmc(unsigned int n, unsigned int m) {
     unsigned int maior = (n > m) ? n : m;
-    unsigned long int i, i_max = n*m;
+    unsigned long int i, i_max = (unsigned long int) n*m;
 
     for (i=maior; i<i_max; i++)
         if ((i%n == 0) && (i%m == 0))
@@ -19,6 +18,5 @@ int main() {
     scanf("%u %u", &n, &m);
 
     printf("O MMC entre %u e %u eh %lu\n", n, m, mmc(n, m));
-    /*printf("%lu", (unsigned long int) n + m);*/
     return 0;
 }
